refactor(cli): split the pipeline runner below the 800-line cap
969 lines against a 1000-line hard gate, and PR #661 adds 40 more to it.
Two layers came out, neither of them orchestration:

- `_asset_fetch.py` — what a memory asks Immich for, given its windows and
  the people in it. `fetch_photos` and `fetch_videos_and_live_photos` are
  the runner's other public entry points; nothing in the runner calls them
  and they call nothing in it.
- `_candidate_pool.py` — videos and photos merged into one pool and then
  filtered. Two of its five functions were already private to the group,
  and the UI's clip pipeline builds its pool through the same door.

Behaviour-neutral: every moved body is byte-identical to main, and the
whole `+` side of the runner's diff is three docstring lines and the
five-line import block.

Part of #685. Unblocks #661.
